Added the doctore revenue for payment

+++ HMSApp/src/main/java/com/hms/entities/Doctor.java
@@ -22,7 +22,7 @@ public class Doctor {
     @Id 
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column
-    private Long doctorId;
+    private int doctorId;
 
     @NotBlank(message = "Doctor name cannot be blank")
     @Size(min = 2, max = 100, message = "Doctor name must be between 2 and 100 characters")
